Restore overallrankings.html page
Now that we've verified the Github workflow works as intended, this
configures the update-rankings.js file to restore the page back to its
original state. The next run at ~4pm will create the bot commit.

Once the Google spreadsheet information is acquired, we can configure
more of the code so that it displays real-time updates.