Add Google spreadsheet auto update
This modifies the update-rankings.js file which runs from a scheduled
Github action every hour to fetch the data from the Google Spreadsheet,
parse it, and auto-update the HTML code.

Currently, the spreadsheet is filled with dummy values from a random
test shop. In the future, when the Shopify store is setup properly, this
will pull the order summary from the exported data and re-render the
rankings page. We might want to look into adding the raw URL link of the
spreadsheet as a Github secret.